chore: move all ReadProjectionConfigs.ReadAs* nested classes to their own files
